Small cleanup, and moved hand.rs to hands/common.rs
