+++ examples/xr.rs
@@ -3,10 +3,10 @@ use bevy::diagnostic::{FrameTimeDiagnosticsPlugin, LogDiagnosticsPlugin};
 use bevy::prelude::*;
 use bevy::transform::components::Transform;
 use bevy_oxr::input::XrInput;
-use bevy_oxr::resources::{XrFrameState, XrInstance, XrSession};
+use bevy_oxr::resources::{XrFrameState, XrSession};
 
 use bevy_oxr::xr_input::actions::XrActionSets;
-use bevy_oxr::xr_input::hand::{HandInputDebugRenderer, OpenXrHandInput};
+use bevy_oxr::xr_input::hands::common::{HandInputDebugRenderer, OpenXrHandInput};
 use bevy_oxr::xr_input::interactions::{
     draw_interaction_gizmos, draw_socket_gizmos, interactions, socket_interactions,
     update_interactable_states, InteractionEvent, Touched, XRDirectInteractor, XRInteractable,
